// distilled-coinbase/src/operations/createWebhookSubscription.ts
export const CreateWebhookSubscriptionInput = Schema.Struct({
  description: Schema.optional(Schema.String),
  eventTypes: Schema.Array(Schema.String),
  isEnabled: Schema.Boolean,
  target: Schema.Struct({
    url: Schema.String,
    headers: Schema.optional(Schema.Record({ key: Schema.String, value: Schema.String })),
  }),
  metadata: Schema.optional(Schema.Record({ key: Schema.String, value: Schema.String })),
  labels: Schema.optional(Schema.Record({ key: Schema.String, value: Schema.String })),
}).pipe(T.Http({ method: "POST", path: "/v2/data/webhooks/subscriptions" }));
export type CreateWebhookSubscriptionInput = typeof CreateWebhookSubscriptionInput.Type;

// Output Schema
export const CreateWebhookSubscriptionOutput = Schema.Struct({
  createdAt: Schema.String,
  description: Schema.optional(Schema.String),
  eventTypes: Schema.Array(Schema.String),
  isEnabled: Schema.Boolean,
  metadata: Schema.optional(Schema.Struct({
    secret: Schema.optional(Schema.String),
  })),
  secret: Schema.String,
  subscriptionId: Schema.String,
  target: Schema.Struct({
    url: Schema.String,
    headers: Schema.optional(Schema.Record({ key: Schema.String, value: Schema.String })),
  }),
  labels: Schema.optional(Schema.Record({ key: Schema.String, value: Schema.String })),
});
export type CreateWebhookSubscriptionOutput = typeof CreateWebhookSubscriptionOutput.Type;
// ...
export const createWebhookSubscription = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: CreateWebhookSubscriptionInput,
  outputSchema: CreateWebhookSubscriptionOutput,
}));